+++ test/common/router/config_impl_test.cc
@@ -2411,6 +2411,7 @@ virtual_hosts:
     route->routeEntry()->grpcTimeoutOffset();
     route->routeEntry()->upgradeMap();
     route->routeEntry()->internalRedirectAction();
+    route->routeEntry()->maxInternalRedirects();
   }
 }